Added the `be_extra_careful` flag to CeperleyMatrix.
This flag, which as defined now used to be true, now defaults to
false.  In particular, the flag should be set to true any time a
negative exponent is going to be applied to a determinant.

When the determinant drops below the `ceperley_determinant_cutoff`, we
recalculate the inverse just to be safe.  If `be_extra_careful` is
true, the inverse is recalculated before reporting the new
determinant.  Otherwise, it is recalculated only if a move is
accepted.